PC-14087 Fix import cycle
