+++ cspell.yaml
@@ -139,6 +139,7 @@ words:
   - sloerrorbudgetstatus
   - sloname
   - slos
+  - slostatusapi
   - sourcetype
   - stablebetaalpha
   - startuml
